Implementation of the extract_MB_jacobi_tridiag function, and declaration and implementation of the jacobi_GB function + Consequential modification of tp_poisson1D_iter.c and lib_poisson1D.h + Addition of extra gnuplot scripts + Addition of assets and data + update of Reponse.md
In extract_MB_jacobi_tridiag, MB is a simplified version of the original GB format, here for jacobi to simply copy the main diagonal.
In jacobi_GB, the calculation of residuals is the same as in richardson_alpha (CBLAS is used), but the calculation of the X solution is different (see reponse.md).
In tp_poisson1D_iter.c, when it is JAC's turn (=1), the number of iterations until convergence and the error in relation to the analytical solution are displayed.
The gnuplot scripts are basic
The assets are used for the reponse.md file, which has been updated.