Implementation of the extract_MB_gauss_seidel_tridiag function, and declaration and implementation of the gauss_seidel_GB function + Consequential modification of tp_poisson1D_iter.c and lib_poisson1D.h + Addition of extra gnuplot scripts + Addition of assets and data + update of Reponse.md
In extract_MB_gauss_seidel_tridiag, MB is a modified version of the original GB format, here we copy the main diagonal and the sub-diagonal (with an inverted sign).
In gauss_seidel_GB, the calculation of residuals is the same as in jacobi_GB and richardson_alpha (CBLAS is used), but the calculation of the X solution is different (see reponse.md).
In tp_poisson1D_iter.c, when it's GS's turn (=2), the number of iterations until convergence and the error in relation to the analytical solution are displayed.
The gnuplot scripts are basic
The assets are used for the reponse.md file, which has been updated.